.github/workflows/scripts/requirements.txt
requests == 2.31.0
GitPython == 3.1.14